start working on shadows

+++ src/utils/Input.cpp
@@ -31,8 +31,8 @@ void Input::setMousePos(Vector2f  pos){
     glfwSetCursorPos(Input::window, pos.x, pos.y);
 }
 void Input::onMouseScroll(GLFWwindow * window, double posX, double posY){
-    mouseScrollOffset.x += posX;
-    mouseScrollOffset.y += posY;
+    mouseScrollOffset.x += (float)posX;
+    mouseScrollOffset.y += (float)posY;
 }
 void Input::onKeyDown(GLFWwindow * window, int key, int scanCode, int action, int mods){
     if(action == GLFW_PRESS){
